+++ src/module/activities/DamageActivity.ts
@@ -37,7 +37,7 @@ let defineMidiDamageActivityClass = (ActivityClass: any) => {
 
     async rollDamage(config, dialog, message) {
       config.midiOptions ??= {};
-      config.midiOptions.fastForwardDamage = game.user?.isGM ? configSettings.gmAutoFastForwardDamage : ["all", "damage"].includes(configSettings.autoFastForward);
+      config.midiOptions.fastForwardDamage ??= game.user?.isGM ? configSettings.gmAutoFastForwardDamage : ["all", "damage"].includes(configSettings.autoFastForward);
       return super.rollDamage(config, dialog, message);
     }